requirements.txt
Flask==2.0.1
scikit-learn==0.24.2
pandas==1.3.0
numpy==1.21.0
python-whois==0.8.0